Move location-related hooks to `context.hooks.location`
