+++ quarkdown-core/src/main/kotlin/com/quarkdown/core/ast/attributes/location/SectionLocationProperty.kt
@@ -6,7 +6,7 @@ import com.quarkdown.core.property.Property
  * [Property] that is assigned to each node that requests its location to be tracked ([LocationTrackableNode]).
  * It contains the node's location in the document, in terms of section indices.
  * @see SectionLocation
- * @see com.quarkdown.core.context.hooks.LocationAwarenessHook for the storing stage
+ * @see com.quarkdown.core.context.hooks.location.LocationAwarenessHook for the storing stage
  */
 data class SectionLocationProperty(
     override val value: SectionLocation,
