+++ quarkdown-core/src/main/kotlin/com/quarkdown/core/flavor/quarkdown/QuarkdownTreeIteratorFactory.kt
@@ -2,10 +2,10 @@ package com.quarkdown.core.flavor.quarkdown
 
 import com.quarkdown.core.ast.iterator.ObservableAstIterator
 import com.quarkdown.core.context.MutableContext
-import com.quarkdown.core.context.hooks.LocationAwareLabelStorerHook
-import com.quarkdown.core.context.hooks.LocationAwarenessHook
 import com.quarkdown.core.context.hooks.MediaStorerHook
 import com.quarkdown.core.context.hooks.TableOfContentsGeneratorHook
+import com.quarkdown.core.context.hooks.location.LocationAwareLabelStorerHook
+import com.quarkdown.core.context.hooks.location.LocationAwarenessHook
 import com.quarkdown.core.flavor.TreeIteratorFactory
 import com.quarkdown.core.flavor.base.BaseMarkdownTreeIteratorFactory
